linux can't serialise unsigned long?

+++ game/ui/battle/battleview.h
@@ -80,7 +80,7 @@ class MotionScannerInfo
 	// 0 = up, then clockwise
     int direction = 0;
 	UString id = "";
-	unsigned long long version = 0;
+	unsigned long version = 0;
 
 	bool operator==(const MotionScannerInfo &other) const;
 	bool operator!=(const MotionScannerInfo &other) const;
